Move a couple of utility functions
